Created MessageList component
